server/requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiomysql==0.0.21
aioredis==1.3.1
aiosmtplib==1.1.6
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
hiredis==2.0.0
idna==3.2
multidict==5.1.0
PyMySQL==0.9.3
PyYAML==5.4.1
selenium==3.141.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
